fixes #2 and renames Table to TableSchema
