+++ src/Database/Schema/BaseSchema.php
@@ -54,19 +54,19 @@ abstract class BaseSchema
      */
     protected function _foreignOnClause($on)
     {
-        if ($on === Table::ACTION_SET_NULL) {
+        if ($on === TableSchema::ACTION_SET_NULL) {
             return 'SET NULL';
         }
-        if ($on === Table::ACTION_SET_DEFAULT) {
+        if ($on === TableSchema::ACTION_SET_DEFAULT) {
             return 'SET DEFAULT';
         }
-        if ($on === Table::ACTION_CASCADE) {
+        if ($on === TableSchema::ACTION_CASCADE) {
             return 'CASCADE';
         }
-        if ($on === Table::ACTION_RESTRICT) {
+        if ($on === TableSchema::ACTION_RESTRICT) {
             return 'RESTRICT';
         }
-        if ($on === Table::ACTION_NO_ACTION) {
+        if ($on === TableSchema::ACTION_NO_ACTION) {
             return 'NO ACTION';
         }
     }
@@ -83,9 +83,9 @@ abstract class BaseSchema
             return strtolower($clause);
         }
         if ($clause === 'NO ACTION') {
-            return Table::ACTION_NO_ACTION;
+            return TableSchema::ACTION_NO_ACTION;
         }
-        return Table::ACTION_SET_NULL;
+        return TableSchema::ACTION_SET_NULL;
     }
 
     /**
@@ -110,10 +110,10 @@ abstract class BaseSchema
     /**
      * Generate the SQL to drop a table.
      *
-     * @param \Cake\Database\Schema\Table $table Table instance
+     * @param \Cake\Database\Schema\TableSchema $table Table instance
      * @return array SQL statements to drop a table.
      */
-    public function dropTableSql(Table $table)
+    public function dropTableSql(TableSchema $table)
     {
         $sql = sprintf(
             'DROP TABLE %s',
@@ -173,102 +173,102 @@ abstract class BaseSchema
     /**
      * Convert field description results into abstract schema fields.
      *
-     * @param \Cake\Database\Schema\Table $table The table object to append fields to.
+     * @param \Cake\Database\Schema\TableSchema $table The table object to append fields to.
      * @param array $row The row data from `describeColumnSql`.
      * @return void
      */
-    abstract public function convertColumnDescription(Table $table, $row);
+    abstract public function convertColumnDescription(TableSchema $table, $row);
 
     /**
      * Convert an index description results into abstract schema indexes or constraints.
      *
-     * @param \Cake\Database\Schema\Table $table The table object to append
+     * @param \Cake\Database\Schema\TableSchema $table The table object to append
      *    an index or constraint to.
      * @param array $row The row data from `describeIndexSql`.
      * @return void
      */
-    abstract public function convertIndexDescription(Table $table, $row);
+    abstract public function convertIndexDescription(TableSchema $table, $row);
 
     /**
      * Convert a foreign key description into constraints on the Table object.
      *
-     * @param \Cake\Database\Schema\Table $table The table object to append
+     * @param \Cake\Database\Schema\TableSchema $table The table object to append
      *    a constraint to.
      * @param array $row The row data from `describeForeignKeySql`.
      * @return void
      */
-    abstract public function convertForeignKeyDescription(Table $table, $row);
+    abstract public function convertForeignKeyDescription(TableSchema $table, $row);
 
     /**
      * Convert options data into table options.
      *
-     * @param \Cake\Database\Schema\Table $table Table instance.
+     * @param \Cake\Database\Schema\TableSchema $table Table instance.
      * @param array $row The row of data.
      * @return void
      */
-    public function convertOptionsDescription(Table $table, $row)
+    public function convertOptionsDescription(TableSchema $table, $row)
     {
     }
 
     /**
      * Generate the SQL to create a table.
      *
-     * @param \Cake\Database\Schema\Table $table Table instance.
+     * @param \Cake\Database\Schema\TableSchema $table Table instance.
      * @param array $columns The columns to go inside the table.
      * @param array $constraints The constraints for the table.
      * @param array $indexes The indexes for the table.
      * @return array SQL statements to create a table.
      */
-    abstract public function createTableSql(Table $table, $columns, $constraints, $indexes);
+    abstract public function createTableSql(TableSchema $table, $columns, $constraints, $indexes);
 
     /**
      * Generate the SQL fragment for a single column in a table.
      *
-     * @param \Cake\Database\Schema\Table $table The table instance the column is in.
+     * @param \Cake\Database\Schema\TableSchema $table The table instance the column is in.
      * @param string $name The name of the column.
      * @return string SQL fragment.
      */
-    abstract public function columnSql(Table $table, $name);
+    abstract public function columnSql(TableSchema $table, $name);
 
     /**
      * Generate the SQL queries needed to add foreign key constraints to the table
      *
-     * @param \Cake\Database\Schema\Table $table The table instance the foreign key constraints are.
+     * @param \Cake\Database\Schema\TableSchema $table The table instance the foreign key constraints are.
      * @return array SQL fragment.
      */
-    abstract public function addConstraintSql(Table $table);
+    abstract public function addConstraintSql(TableSchema $table);
 
     /**
      * Generate the SQL queries needed to drop foreign key constraints from the table
      *
-     * @param \Cake\Database\Schema\Table $table The table instance the foreign key constraints are.
+     * @param \Cake\Database\Schema\TableSchema $table The table instance the foreign key constraints are.
      * @return array SQL fragment.
      */
-    abstract public function dropConstraintSql(Table $table);
+    abstract public function dropConstraintSql(TableSchema $table);
 
     /**
      * Generate the SQL fragments for defining table constraints.
      *
-     * @param \Cake\Database\Schema\Table $table The table instance the column is in.
+     * @param \Cake\Database\Schema\TableSchema $table The table instance the column is in.
      * @param string $name The name of the column.
      * @return string SQL fragment.
      */
-    abstract public function constraintSql(Table $table, $name);
+    abstract public function constraintSql(TableSchema $table, $name);
 
     /**
      * Generate the SQL fragment for a single index in a table.
      *
-     * @param \Cake\Database\Schema\Table $table The table object the column is in.
+     * @param \Cake\Database\Schema\TableSchema $table The table object the column is in.
      * @param string $name The name of the column.
      * @return string SQL fragment.
      */
-    abstract public function indexSql(Table $table, $name);
+    abstract public function indexSql(TableSchema $table, $name);
 
     /**
      * Generate the SQL to truncate a table.
      *
-     * @param \Cake\Database\Schema\Table $table Table instance.
+     * @param \Cake\Database\Schema\TableSchema $table Table instance.
      * @return array SQL statements to truncate a table.
      */
-    abstract public function truncateTableSql(Table $table);
+    abstract public function truncateTableSql(TableSchema $table);
 }
