+++ src/Database/Schema/Collection.php
@@ -82,7 +82,7 @@ class Collection
      *
      * @param string $name The name of the table to describe.
      * @param array $options The options to use, see above.
-     * @return \Cake\Database\Schema\Table Object with column metadata.
+     * @return \Cake\Database\Schema\TableSchema Object with column metadata.
      * @throws \Cake\Database\Exception when table cannot be described.
      */
     public function describe($name, array $options = [])
@@ -91,7 +91,7 @@ class Collection
         if (strpos($name, '.')) {
             list($config['schema'], $name) = explode('.', $name);
         }
-        $table = new Table($name);
+        $table = new TableSchema($name);
 
         $this->_reflect('Column', $name, $config, $table);
         if (count($table->columns()) === 0) {
@@ -111,7 +111,7 @@ class Collection
      * @param string $stage The stage name.
      * @param string $name The table name.
      * @param array $config The config data.
-     * @param \Cake\Database\Schema\Table $table The table instance
+     * @param \Cake\Database\Schema\TableSchema $table The table instance
      * @return void
      * @throws \Cake\Database\Exception on query failure.
      */
