+++ src/ORM/Table.php
@@ -17,7 +17,7 @@ namespace Cake\ORM;
 use ArrayObject;
 use BadMethodCallException;
 use Cake\Core\App;
-use Cake\Database\Schema\Table as Schema;
+use Cake\Database\Schema\TableSchema;
 use Cake\Database\Type;
 use Cake\Datasource\ConnectionInterface;
 use Cake\Datasource\EntityInterface;
@@ -174,7 +174,7 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
     /**
      * The schema object containing a description of this table fields
      *
-     * @var \Cake\Database\Schema\Table
+     * @var \Cake\Database\Schema\TableSchema
      */
     protected $_schema;
 
@@ -230,7 +230,7 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
      * - connection: The connection instance to use
      * - entityClass: The fully namespaced class name of the entity class that will
      *   represent rows in this table.
-     * - schema: A \Cake\Database\Schema\Table object or an array that can be
+     * - schema: A \Cake\Database\Schema\TableSchema object or an array that can be
      *   passed to it.
      * - eventManager: An instance of an event manager to use for internal events
      * - behaviors: A BehaviorRegistry. Generally not used outside of tests.
@@ -410,14 +410,14 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
     /**
      * Returns the schema table object describing this table's properties.
      *
-     * If an \Cake\Database\Schema\Table is passed, it will be used for this table
+     * If an \Cake\Database\Schema\TableSchema is passed, it will be used for this table
      * instead of the default one.
      *
-     * If an array is passed, a new \Cake\Database\Schema\Table will be constructed
+     * If an array is passed, a new \Cake\Database\Schema\TableSchema will be constructed
      * out of it and used as the schema for this table.
      *
-     * @param array|\Cake\Database\Schema\Table|null $schema New schema to be used for this table
-     * @return \Cake\Database\Schema\Table
+     * @param array|\Cake\Database\Schema\TableSchema|null $schema New schema to be used for this table
+     * @return \Cake\Database\Schema\TableSchema
      */
     public function schema($schema = null)
     {
@@ -440,7 +440,7 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
                 unset($schema['_constraints']);
             }
 
-            $schema = new Schema($this->table(), $schema);
+            $schema = new TableSchema($this->table(), $schema);
 
             foreach ($constraints as $name => $value) {
                 $schema->addConstraint($name, $value);
@@ -460,17 +460,17 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
      * ### Example:
      *
      * ```
-     * protected function _initializeSchema(\Cake\Database\Schema\Table $table) {
+     * protected function _initializeSchema(\Cake\Database\Schema\TableSchema $table) {
      *  $table->columnType('preferences', 'json');
      *  return $table;
      * }
      * ```
      *
-     * @param \Cake\Database\Schema\Table $table The table definition fetched from database.
-     * @return \Cake\Database\Schema\Table the altered schema
+     * @param \Cake\Database\Schema\TableSchema $table The table definition fetched from database.
+     * @return \Cake\Database\Schema\TableSchema the altered schema
      * @api
      */
-    protected function _initializeSchema(Schema $table)
+    protected function _initializeSchema(TableSchema $table)
     {
         return $table;
     }
@@ -479,7 +479,7 @@ class Table implements RepositoryInterface, EventListenerInterface, EventDispatc
      * Test to see if a Table has a specific field/column.
      *
      * Delegates to the schema object and checks for column presence
-     * using the Schema\Table instance.
+     * using the \Cake\Database\Schema\TableSchema instance.
      *
      * @param string $field The field to check for.
      * @return bool True if the field exists, false if it does not.
