+++ tests/Fixture/TagsFixture.php
@@ -14,7 +14,7 @@
  */
 namespace Cake\Test\Fixture;
 
-use Cake\Database\Schema\Table;
+use Cake\Database\Schema\TableSchema;
 use Cake\TestSuite\Fixture\TestFixture;
 
 /**
@@ -32,7 +32,7 @@ class TagsFixture extends TestFixture
     public $fields = [
         'id' => ['type' => 'integer', 'null' => false],
         'name' => ['type' => 'string', 'null' => false],
-        'description' => ['type' => 'text', 'length' => Table::LENGTH_MEDIUM],
+        'description' => ['type' => 'text', 'length' => TableSchema::LENGTH_MEDIUM],
         '_constraints' => ['primary' => ['type' => 'primary', 'columns' => ['id']]]
     ];
