+++ tests/TestCase/Database/Schema/MysqlSchemaTest.php
@@ -17,7 +17,7 @@ namespace Cake\Test\TestCase\Database\Schema;
 use Cake\Core\Configure;
 use Cake\Database\Schema\Collection as SchemaCollection;
 use Cake\Database\Schema\MysqlSchema;
-use Cake\Database\Schema\Table;
+use Cake\Database\Schema\TableSchema;
 use Cake\Datasource\ConnectionManager;
 use Cake\TestSuite\TestCase;
 
@@ -104,19 +104,19 @@ class MysqlSchemaTest extends TestCase
             ],
             [
                 'TINYTEXT',
-                ['type' => 'text', 'length' => Table::LENGTH_TINY]
+                ['type' => 'text', 'length' => TableSchema::LENGTH_TINY]
             ],
             [
                 'MEDIUMTEXT',
-                ['type' => 'text', 'length' => Table::LENGTH_MEDIUM]
+                ['type' => 'text', 'length' => TableSchema::LENGTH_MEDIUM]
             ],
             [
                 'LONGTEXT',
-                ['type' => 'text', 'length' => Table::LENGTH_LONG]
+                ['type' => 'text', 'length' => TableSchema::LENGTH_LONG]
             ],
             [
                 'TINYBLOB',
-                ['type' => 'binary', 'length' => Table::LENGTH_TINY]
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_TINY]
             ],
             [
                 'BLOB',
@@ -124,11 +124,11 @@ class MysqlSchemaTest extends TestCase
             ],
             [
                 'MEDIUMBLOB',
-                ['type' => 'binary', 'length' => Table::LENGTH_MEDIUM]
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_MEDIUM]
             ],
             [
                 'LONGBLOB',
-                ['type' => 'binary', 'length' => Table::LENGTH_LONG]
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_LONG]
             ],
             [
                 'FLOAT',
@@ -195,7 +195,7 @@ class MysqlSchemaTest extends TestCase
         $driver = $this->getMock('Cake\Database\Driver\Mysql');
         $dialect = new MysqlSchema($driver);
 
-        $table = $this->getMock('Cake\Database\Schema\Table', [], ['table']);
+        $table = $this->getMock('Cake\Database\Schema\TableSchema', [], ['table']);
         $table->expects($this->at(0))->method('addColumn')->with('field', $expected);
 
         $dialect->convertColumnDescription($table, $field);
@@ -270,7 +270,7 @@ SQL;
 
         $schema = new SchemaCollection($connection);
         $result = $schema->describe('schema_articles');
-        $this->assertInstanceOf('Cake\Database\Schema\Table', $result);
+        $this->assertInstanceOf('Cake\Database\Schema\TableSchema', $result);
         $expected = [
             'id' => [
                 'type' => 'biginteger',
@@ -356,7 +356,7 @@ SQL;
 
         $schema = new SchemaCollection($connection);
         $result = $schema->describe('schema_articles');
-        $this->assertInstanceOf('Cake\Database\Schema\Table', $result);
+        $this->assertInstanceOf('Cake\Database\Schema\TableSchema', $result);
 
         $this->assertCount(3, $result->constraints());
         $expected = [
@@ -488,17 +488,17 @@ SQL;
             ],
             [
                 'body',
-                ['type' => 'text', 'length' => Table::LENGTH_TINY, 'null' => false],
+                ['type' => 'text', 'length' => TableSchema::LENGTH_TINY, 'null' => false],
                 '`body` TINYTEXT NOT NULL'
             ],
             [
                 'body',
-                ['type' => 'text', 'length' => Table::LENGTH_MEDIUM, 'null' => false],
+                ['type' => 'text', 'length' => TableSchema::LENGTH_MEDIUM, 'null' => false],
                 '`body` MEDIUMTEXT NOT NULL'
             ],
             [
                 'body',
-                ['type' => 'text', 'length' => Table::LENGTH_LONG, 'null' => false],
+                ['type' => 'text', 'length' => TableSchema::LENGTH_LONG, 'null' => false],
                 '`body` LONGTEXT NOT NULL'
             ],
             // Blob / binary
@@ -509,17 +509,17 @@ SQL;
             ],
             [
                 'body',
-                ['type' => 'binary', 'length' => Table::LENGTH_TINY, 'null' => false],
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_TINY, 'null' => false],
                 '`body` TINYBLOB NOT NULL'
             ],
             [
                 'body',
-                ['type' => 'binary', 'length' => Table::LENGTH_MEDIUM, 'null' => false],
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_MEDIUM, 'null' => false],
                 '`body` MEDIUMBLOB NOT NULL'
             ],
             [
                 'body',
-                ['type' => 'binary', 'length' => Table::LENGTH_LONG, 'null' => false],
+                ['type' => 'binary', 'length' => TableSchema::LENGTH_LONG, 'null' => false],
                 '`body` LONGBLOB NOT NULL'
             ],
             // Integers
@@ -643,7 +643,7 @@ SQL;
         $driver = $this->_getMockedDriver();
         $schema = new MysqlSchema($driver);
 
-        $table = (new Table('articles'))->addColumn($name, $data);
+        $table = (new TableSchema('articles'))->addColumn($name, $data);
         $this->assertEquals($expected, $schema->columnSql($table, $name));
     }
 
@@ -717,7 +717,7 @@ SQL;
         $driver = $this->_getMockedDriver();
         $schema = new MysqlSchema($driver);
 
-        $table = (new Table('articles'))->addColumn('title', [
+        $table = (new TableSchema('articles'))->addColumn('title', [
             'type' => 'string',
             'length' => 255
         ])->addColumn('author_id', [
@@ -758,7 +758,7 @@ SQL;
         $driver = $this->_getMockedDriver();
         $schema = new MysqlSchema($driver);
 
-        $table = (new Table('articles'))->addColumn('title', [
+        $table = (new TableSchema('articles'))->addColumn('title', [
             'type' => 'string',
             'length' => 255
         ])->addColumn('author_id', [
@@ -780,7 +780,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = (new Table('posts'))
+        $table = (new TableSchema('posts'))
             ->addColumn('author_id', [
                 'type' => 'integer',
                 'null' => false
@@ -829,7 +829,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = (new Table('posts'))
+        $table = (new TableSchema('posts'))
             ->addColumn('author_id', [
                 'type' => 'integer',
                 'null' => false
@@ -876,7 +876,7 @@ SQL;
         $driver = $this->_getMockedDriver();
         $schema = new MysqlSchema($driver);
 
-        $table = new Table('articles');
+        $table = new TableSchema('articles');
         $table->addColumn('id', [
                 'type' => 'integer',
                 'null' => false,
@@ -888,7 +888,7 @@ SQL;
         $result = $schema->columnSql($table, 'id');
         $this->assertEquals($result, '`id` INTEGER NOT NULL AUTO_INCREMENT');
 
-        $table = new Table('articles');
+        $table = new TableSchema('articles');
         $table->addColumn('id', [
                 'type' => 'biginteger',
                 'null' => false
@@ -913,7 +913,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = (new Table('posts'))->addColumn('id', [
+        $table = (new TableSchema('posts'))->addColumn('id', [
                 'type' => 'integer',
                 'null' => false
             ])
@@ -962,7 +962,7 @@ SQL;
         $connection = $this->getMock('Cake\Database\Connection', [], [], '', false);
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
-        $table = (new Table('schema_articles'))->addColumn('id', [
+        $table = (new TableSchema('schema_articles'))->addColumn('id', [
             'type' => 'integer',
             'null' => false
         ]);
@@ -983,7 +983,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = (new Table('articles_tags'))
+        $table = (new TableSchema('articles_tags'))
             ->addColumn('article_id', [
                 'type' => 'integer',
                 'null' => false
@@ -1008,7 +1008,7 @@ SQL;
         $this->assertCount(1, $result);
         $this->assertTextEquals($expected, $result[0]);
 
-        $table = (new Table('composite_key'))
+        $table = (new TableSchema('composite_key'))
             ->addColumn('id', [
                 'type' => 'integer',
                 'null' => false,
@@ -1047,7 +1047,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = new Table('articles');
+        $table = new TableSchema('articles');
         $result = $table->dropSql($connection);
         $this->assertCount(1, $result);
         $this->assertEquals('DROP TABLE `articles`', $result[0]);
@@ -1065,7 +1065,7 @@ SQL;
         $connection->expects($this->any())->method('driver')
             ->will($this->returnValue($driver));
 
-        $table = new Table('articles');
+        $table = new TableSchema('articles');
         $result = $table->truncateSql($connection);
         $this->assertCount(1, $result);
         $this->assertEquals('TRUNCATE TABLE `articles`', $result[0]);
